Conflicts:
	src/maidsafe/client/maidstoremanager.h

+++ src/maidsafe/client/clientcontroller.cc
@@ -633,14 +633,6 @@ bool ClientController::ValidateUser(const std::string &password) {
                              &ClientController::ClearStaleMessages, this);
   }
 
-//  // CHANGE CONNECTION STATUS
-//  int connection_status(1);
-//  int n = ChangeConnectionStatus(connection_status);
-//  if (n != 0) {
-//    // Alert for BP problems
-//  }
-//  ss_->SetConnectionStatus(connection_status);
-
   logged_in_ = true;
   return true;
 }
@@ -693,13 +685,6 @@ bool ClientController::Logout() {
   }
   logging_out_ = true;
   clear_messages_thread_.join();
-
-//  int connection_status(0);
-//  int n = ChangeConnectionStatus(connection_status);
-//  if (n != 0) {
-//    // Alert for BP problems
-//  }
-//  ss_->SetConnectionStatus(connection_status);
 
   int result = SaveSession();
   if (result != kSuccess) {
@@ -1585,7 +1570,7 @@ int ClientController::SendInstantFile(std::string *filename,
 void ClientController::onInstantMessage(const std::string &message,
                                         const boost::uint32_t&,
                                         const boost::int16_t&,
-                                        const float&) {
+                                        const double&) {
   GenericPacket gp;
   if (!gp.ParseFromString(message)) {
     return;
